# cmd/agents-server/internal/docs/swagger.yaml
components:
  schemas:
...
    handler.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/protocol.APIError'
      type: object
...
    handler.UserLabel:
      properties:
        email:
          type: string
        id:
          type: string
        name:
          type: string
      type: object
    handler.UserPatchRequest:
      properties:
        disabled:
          type: boolean
        role:
          type: string
      type: object
...
    protocol.APIError:
      properties:
        code:
          type: string
        message:
          type: string
      type: object
...
    protocol.AuthSession:
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/protocol.UserInfo'
      type: object
    protocol.PatCreated:
      properties:
        pat:
          $ref: '#/components/schemas/protocol.PatView'
        token:
          type: string
      type: object
    protocol.PatView:
      properties:
        created_at:
          type: string
        expires_at:
          type: string
        id:
          type: string
        last_used_at:
          type: string
        name:
          type: string
      type: object
    protocol.UserInfo:
      properties:
        avatar_url:
          description: AvatarURL is the login provider's picture, when it gave one.
          type: string
        email:
          type: string
        id:
          type: string
        name:
          type: string
        role:
          type: string
      type: object
...
    store.AuditEvent:
      properties:
        action:
          description: |-
            Action is "METHOD /route/pattern" for REST, or a dotted name for the
            explicit events (ws.run.create, ws.approval, terminal.open).
          type: string
        actor_email:
          type: string
        actor_id:
          description: |-
            ActorID/ActorEmail identify the caller; the email is a snapshot so the
            line stays readable after the account is gone.
          type: string
        client_ip:
          type: string
        created_at:
          type: string
        detail:
          type: string
        id:
          type: string
        resource:
          type: string
      type: object
...
    store.User:
      properties:
        avatar_url:
          description: |-
            AvatarURL is the provider's picture URL, carried by /auth/me and loaded
            by the browser; the CSP admits the configured providers' image hosts.
          type: string
        created_at:
          type: string
        disabled_at:
          description: |-
            DisabledAt, when set, is when an admin switched the account off: no
            credential of theirs authenticates until it is cleared.
          type: string
        email:
          description: lowercased; unique via idx_users_email
          type: string
        id:
          type: string
        last_login_at:
          type: string
        name:
          type: string
        role:
          description: RoleAdmin | RoleMember
          type: string
        updated_at:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Value is "Bearer <token>" — the token printed at server startup
        (or set via --token).
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 'REST API of agents-server, the web server over the agents-go SDK:
    sessions, runs, agents, MCP servers, memories, settings, skills, providers, workflows,
    triggers, guardrails, sandboxes and projects. A run starts over REST (POST /sessions/{id}/runs)
    or the WebSocket protocol (docs/reference/protocol.md) and streams over either;
    this spec covers the HTTP surface.'
  title: agents-server API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /auth/audit:
    get:
      parameters:
      - description: Page size (default 100, max 500)
        in: query
        name: limit
        schema:
          type: integer
      - description: An event id; the page before it
        in: query
        name: before
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.AuditEvent'
                type: array
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Audit log (admin)
      tags:
      - auth
  /auth/check:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Validate the presented credential
      tags:
      - auth
...
  /auth/exchange:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/protocol.AuthSession'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: unknown, used, or expired code
      summary: Exchange the one-time login code for a session token
      tags:
      - auth
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: malformed body, or token login disabled (OAuth mode)
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
      summary: Token login
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "204":
          description: signed out
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Sign out
      tags:
      - auth
  /auth/me:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/protocol.UserInfo'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /auth/oauth/{provider}/callback:
    get:
      parameters:
      - description: Provider name
        in: path
        name: provider
        required: true
        schema:
          type: string
      - description: Authorize round-trip state
        in: query
        name: state
        required: true
        schema:
          type: string
      - description: Authorization code
        in: query
        name: code
        schema:
          type: string
      - description: The provider's error (a cancelled consent)
        in: query
        name: error
        schema:
          type: string
      responses:
        "302":
          description: redirect into the SPA
      summary: OAuth provider callback
      tags:
      - auth
  /auth/oauth/{provider}/start:
    get:
      parameters:
      - description: Provider name (e.g. google)
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "302":
          description: redirect to the provider's authorize URL
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: unknown provider (or token mode)
      summary: Begin an OAuth login
      tags:
      - auth
  /auth/tokens:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/protocol.PatView'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: token mode
      security:
      - BearerAuth: []
      summary: List personal access tokens
      tags:
      - auth
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/protocol.PatCreated'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
      security:
      - BearerAuth: []
      summary: Create a personal access token
      tags:
      - auth
  /auth/tokens/{id}:
    delete:
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: revoked
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Revoke a personal access token
      tags:
      - auth
  /auth/user-labels:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/handler.UserLabel'
                type: array
          description: OK
      security:
      - BearerAuth: []
      summary: List user labels (id, name, email) for owner display
      tags:
      - auth
  /auth/users:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.User'
                type: array
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
      security:
      - BearerAuth: []
      summary: List users (admin)
      tags:
      - auth
  /auth/users/{id}:
    patch:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.UserPatchRequest'
                description: What to change
                summary: body
        description: What to change
        required: true
      responses:
        "204":
          description: changed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: that would leave no admin
      security:
      - BearerAuth: []
      summary: Change a user's role or disable them (admin)
      tags:
      - auth
  /auth/users/{id}/tokens:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: revoked
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Revoke every token of a user (admin)
      tags:
      - auth
...
servers:
- url: /api/v1